% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_talk_index.R
\name{add_talk_index}
\alias{add_talk_index}
\title{Add Talk Index Page}
\usage{
add_talk_index(
  source = "slides.qmd",
  talk_metadata = yaml::read_yaml("_talk.yml"),
  data = list(title = talk_metadata$title, talk = list(date = talk_metadata$talk$date,
    venue = talk_metadata$talk$venue, location = talk_metadata$talk$location, format =
    talk_metadata$talk$format), repo = list(remote = usethis::git_remotes()$origin, local
    = fs::path_file(usethis::proj_get())), slides = list(pdf = fs::path_ext_set(source,
    ".pdf"), html = fs::path_ext_set(source, ".html"), source = source)),
  save_as = "index.qmd",
  template = "talk-index.qmd",
  ...
)
}
\arguments{
\item{data}{A list of data passed to the template.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{template}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}

\item{...}{Additional optional arguments to update the default data list.}

\item{slides}{A character string specifying the source file for the slides.}
}
\value{
None
}
\description{
This function adds an index page to a talk repository using metadata from the active repository.
It is a wrapper around usethis::use_template('talk-index.qmd', package = 'cynthiahqy')
The template can be found in `cynthiahqy/inst/templates`.
}
\examples{
\dontrun{
add_talk_index()
}
}
