% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_talk_index.R
\name{clip_talk_index}
\alias{clip_talk_index}
\title{Generate Talk YAML Header}
\usage{
clip_talk_index(
  source = "slides.qmd",
  talk_metadata = yaml::read_yaml("_talk.yml"),
  data = list(title = talk_metadata$title, talk = list(date = talk_metadata$talk$date,
    venue = talk_metadata$talk$venue, location = talk_metadata$talk$location, format =
    talk_metadata$talk$format), repo = list(remote = usethis::git_remotes()$origin, local
    = fs::path_file(usethis::proj_get())), slides = list(pdf = fs::path_ext_set(source,
    ".pdf"), html = fs::path_ext_set(source, ".html"), source = source)),
  ...
)
}
\arguments{
\item{source}{A character string specifying the source file for the slides.}

\item{talk_metadata}{A list object with metadata related to the talk.}

\item{data}{A list of data passed to the template.}

\item{...}{Additional optional arguments to update the default data list.}
}
\value{
None
}
\description{
This function generates the YAML header for a talk page and
adds it to the clipboard
}
\examples{
\dontrun{
add_talk_index()
}
}
