% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_blog_post.R
\name{new_blog_post}
\alias{new_blog_post}
\title{Create a New Blog Post}
\usage{
new_blog_post(
  slug,
  save_as = fs::path(slug, "index.qmd"),
  template = "blog-post.qmd",
  data = list(title = "", description = "", date = Sys.Date(), draft = "true", categories
    = list()),
  ...
)
}
\arguments{
\item{slug}{A character string representing the slug for the blog post.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{template}{Path to template file relative to \verb{templates/} directory
within \code{package}; see details.}

\item{data}{A list of data passed to the template.}

\item{...}{Additional optional arguments to update the default data list.}
}
\value{
None
}
\description{
This function creates a new blog post with the specified parameters.
It is a light wrapper around usethis::use_template('blog-post.qmd', package = 'cynthiahqy')
The template 'blog-post.qmd' can be found in `cynthiahqy/inst/templates`.
}
\examples{
\dontrun{
# Note: running this will write `my-third-post/index.qmd` to your working directory
new_blog_post("my-third-post", categories = list("quarto", "git"))
}
}
